requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Brotli==1.1.0
cffi==1.17.1
crispy-tailwind==1.0.3
cssselect2==0.7.0
django==4.2.23
django-cors-headers==4.4.0
django-crispy-forms==2.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.57.0
gunicorn==23.0.0
html5lib==1.1
packaging==25.0
pillow==10.4.0
pycparser==2.22
pydyf==0.8.0
PyJWT==2.9.0
pyphen==0.16.0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
typing-extensions==4.13.2
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3.post1
